# Multithread cp for directories