I don't know whether multithreading like this in Base is allowed, but this appears to speed up `cp` on directories on my setup significantly (6-thread SSD macbook, MacOS)

Master
```julia
julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  6.647516 seconds (287.42 k allocations: 33.990 MiB)

julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  7.410636 seconds (287.42 k allocations: 33.990 MiB, 0.43% gc time)

julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  6.996436 seconds (287.43 k allocations: 33.991 MiB, 1.87% gc time)
```

This PR
```julia
julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  1.514411 seconds (409.63 k allocations: 59.738 MiB, 0.36% gc time)

julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  1.621347 seconds (409.63 k allocations: 59.738 MiB, 0.35% gc time)

julia> @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true);
  1.682619 seconds (409.61 k allocations: 59.738 MiB, 0.55% gc time)


# Parity check

julia> d = mktempdir()
"/var/folders/_6/1yf6sj0950vcg4t91m9ltb5w0000gn/T/jl_bY80nJ"

julia> @time cp("/Users/ian/.julia/registries/General", d, force=true);
  1.688429 seconds (409.65 k allocations: 59.739 MiB, 0.19% gc time)

julia> Pkg.GitTools.tree_hash("/Users/ian/.julia/registries/General") == Pkg.GitTools.tree_hash(d)
true
```