@IanButterworth 2021-02-14
The typical load-balanced `ntasks=Threads.nthreads()+1` default is backed up by a pretty repeatable minima at `ntasks=7` on a 6 thread system

```julia
julia> for ntasks in 1:20
           t = @elapsed cp("/Users/ian/.julia/registries/General", mktempdir(), force=true,  ntasks=ntasks)
           @show ntasks, t
       end

(ntasks, t) = (1, 5.648116088)
(ntasks, t) = (2, 2.312361291)
(ntasks, t) = (3, 2.087711314)
(ntasks, t) = (4, 1.805751717)
(ntasks, t) = (5, 1.880124248)
(ntasks, t) = (6, 1.830967428)
(ntasks, t) = (7, 1.798523184)
(ntasks, t) = (8, 1.889625656)
(ntasks, t) = (9, 2.081643231)
(ntasks, t) = (10, 2.062828056)
(ntasks, t) = (11, 1.95200327)
(ntasks, t) = (12, 1.954158919)
(ntasks, t) = (13, 1.987198201)
(ntasks, t) = (14, 2.155334364)
(ntasks, t) = (15, 2.048653877)
(ntasks, t) = (16, 2.170309276)
(ntasks, t) = (17, 2.155259291)
(ntasks, t) = (18, 2.246891216)
(ntasks, t) = (19, 2.180693019)
(ntasks, t) = (20, 2.14442822)
```