@IanButterworth 2021-02-14
In a worst-case, if users already spawn a `cp` action per-thread, the behavior is still performant.

i.e. here there are 36 concurrent `sendfile` tasks, and the average time per `cp` is ~1.6s (9.4s/6)
```julia
julia> Threads.@threads for _ in 1:Threads.nthreads()
           @time cp("/Users/ian/.julia/registries/General", mktempdir(), force=true)
       end
  9.087354 seconds (2.28 M allocations: 332.051 MiB, 0.45% gc time)
  9.407884 seconds (2.37 M allocations: 345.358 MiB, 0.43% gc time)
  9.562321 seconds (2.41 M allocations: 351.945 MiB, 0.42% gc time)
  9.618445 seconds (2.43 M allocations: 354.279 MiB, 0.42% gc time)
  9.629988 seconds (2.43 M allocations: 354.760 MiB, 0.42% gc time)
  9.774446 seconds (2.46 M allocations: 358.397 MiB, 0.88% gc time)
```