@vtjnash 2021-02-15
Seems worthwhile then. At the lowest layer, the last parameter to libuv `uv_fs_*` functions is whether to run this asynchronously (mostly with threads, since the existence of an async kernel API is still quite new), which is currently turned off. So we might want to toggle it there instead?