@IanButterworth 2021-02-15
If we did it there, would we still need to semaphore? My thinking was that semaphoring and controlling tasks at the julia level would be necessary in case the directory had a huge number of files, so spawning a limited number of tasks at a time would protect against overflow?