@IanButterworth 2021-02-15
Results from an ubuntu VM with 4 threads
Master
```
julia> @time cp("/home/parallels/.julia/registries/General", mktempdir(), force=true);
  0.845588 seconds (288.64 k allocations: 33.191 MiB)

julia> @time cp("/home/parallels/.julia/registries/General", mktempdir(), force=true);
  0.931749 seconds (288.64 k allocations: 33.191 MiB, 3.83% gc time)
```

This PR
```
julia> @time cp("/home/parallels/.julia/registries/General", mktempdir(), force=true);
  0.537254 seconds (411.42 k allocations: 42.860 MiB)

julia> @time cp("/home/parallels/.julia/registries/General", mktempdir(), force=true);
  0.453760 seconds (411.49 k allocations: 42.862 MiB, 2.75% gc time)
```