@vtjnash 2021-02-15
Yes, it could be used without spawning Tasks, letting it manage the queue internally