@IanButterworth 2021-02-27
> At the lowest layer, the last parameter to libuv uv_fs_* functions is whether to run this asynchronously

I took a look and the last param is a callback function. I couldn't figure out how to reorganize things to handle and synchronize the callbacks