@IanButterworth 2021-04-08
@vtjnash just checking this implementation isn't worth merging? Perhaps the other could come later?