@vtjnash 2021-04-09
Yes, it generally would be better to use the IO load balancer, rather than simultaneously oversubscribing the IO queue, and blocking the CPU queue.