@IanButterworth 2021-04-10
I took another look but still don't understand how it would work.

If the outer `sendfile` currently opens and closes the `src` and `dst`, is your suggestion that the close goes into the callback for `jl_fs_sendfile` within the inner `sendfile`?
But the inner `sendfile` seems to send the file in chunks, so the callback would have to be aware if it was on the last chunk, which could be nontrivial if the chunks ended up being finished out of order given async.

Also, couldn't that approach result in too many files open?

Outer
https://github.com/JuliaLang/julia/blob/fc69c9aef38a1b53324abe9a715514f9716d270b/base/file.jl#L963-L983

Inner
https://github.com/JuliaLang/julia/blob/fc69c9aef38a1b53324abe9a715514f9716d270b/base/filesystem.jl#L116-L129