@IanButterworth 2022-11-01
Closing this as stale, and it seems it's better suited to someone more familiar with libuv